Refactored Edit Pages

+++ lib/views/edit_page/edit_street.dart
@@ -10,6 +10,7 @@ import 'package:churchdata/models/data_dialog.dart';
 import 'package:churchdata/models/search_filters.dart';
 import 'package:churchdata/models/list.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:firebase_auth/firebase_auth.dart' as auth;
 import 'package:firebase_crashlytics/firebase_crashlytics.dart'
     if (dart.library.html) 'package:churchdata/FirebaseWeb.dart' hide User;
@@ -26,14 +27,9 @@ class EditStreet extends StatefulWidget {
 }
 
 class _EditStreetState extends State<EditStreet> {
-  List<FocusNode> foci = [FocusNode(), FocusNode(), FocusNode(), FocusNode()];
-
   Map<String, dynamic> old;
   GlobalKey<FormState> form = GlobalKey<FormState>();
-
   Street street;
-
-  String areasSearchFilter = '';
 
   @override
   Widget build(BuildContext context) {
@@ -47,163 +43,143 @@ class _EditStreetState extends State<EditStreet> {
         key: form,
         child: Padding(
           padding: EdgeInsets.all(5),
-          child: SingleChildScrollView(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: <Widget>[
-                Container(
-                  padding: EdgeInsets.symmetric(vertical: 4.0),
-                  child: TextFormField(
+          child: ListView(
+            children: <Widget>[
+              Container(
+                padding: EdgeInsets.symmetric(vertical: 4.0),
+                child: TextFormField(
+                  decoration: InputDecoration(
+                    labelText: 'اسم الشارع',
+                    border: OutlineInputBorder(
+                      borderSide:
+                          BorderSide(color: Theme.of(context).primaryColor),
+                    ),
+                  ),
+                  textInputAction: TextInputAction.next,
+                  onFieldSubmitted: (_) => FocusScope.of(context).nextFocus(),
+                  initialValue: street.name,
+                  onChanged: nameChanged,
+                  validator: (value) {
+                    if (value.isEmpty) {
+                      return 'هذا الحقل مطلوب';
+                    }
+                    return null;
+                  },
+                ),
+              ),
+              Container(
+                padding: EdgeInsets.symmetric(vertical: 4.0),
+                child: InkWell(
+                  onTap: selectArea,
+                  child: InputDecorator(
                     decoration: InputDecoration(
-                      labelText: 'اسم الشارع',
+                      labelText: 'داخل منطقة',
                       border: OutlineInputBorder(
                         borderSide:
                             BorderSide(color: Theme.of(context).primaryColor),
                       ),
                     ),
-                    focusNode: foci[0],
-                    textInputAction: TextInputAction.next,
-                    onFieldSubmitted: (_) {
-                      foci[1].requestFocus();
-                      selectArea();
-                    },
-                    initialValue: street.name,
-                    onChanged: nameChanged,
-                    validator: (value) {
-                      if (value.isEmpty) {
-                        return 'هذا الحقل مطلوب';
-                      }
-                      return null;
-                    },
+                    child: FutureBuilder(
+                        future:
+                            street.areaId == null ? null : street.getAreaName(),
+                        builder: (con, data) {
+                          if (data.connectionState == ConnectionState.done) {
+                            return Text(data.data);
+                          } else if (data.connectionState ==
+                              ConnectionState.waiting) {
+                            return LinearProgressIndicator();
+                          } else {
+                            return Text('لا يوجد');
+                          }
+                        }),
                   ),
                 ),
-                Container(
-                  padding: EdgeInsets.symmetric(vertical: 4.0),
-                  child: Focus(
-                    focusNode: foci[1],
-                    child: InkWell(
-                      onTap: selectArea,
-                      child: InputDecorator(
-                        decoration: InputDecoration(
-                          labelText: 'داخل منطقة',
-                          border: OutlineInputBorder(
-                            borderSide: BorderSide(
-                                color: Theme.of(context).primaryColor),
+              ),
+              ElevatedButton.icon(
+                  icon: Icon(
+                    const IconData(0xe568, fontFamily: 'MaterialIconsR'),
+                  ),
+                  label: Text('تعديل مكان الشارع على الخريطة'),
+                  onPressed: () async {
+                    var oldPoints = street.locationPoints?.sublist(0);
+                    var rslt = await Navigator.of(context).push(
+                      MaterialPageRoute(
+                        builder: (context) => Scaffold(
+                          appBar: AppBar(
+                            actions: <Widget>[
+                              IconButton(
+                                icon: Icon(Icons.done),
+                                onPressed: () => Navigator.pop(context, true),
+                                tooltip: 'حفظ',
+                              ),
+                              IconButton(
+                                icon: Icon(Icons.delete),
+                                onPressed: () => Navigator.pop(context, false),
+                                tooltip: 'حذف التحديد',
+                              )
+                            ],
+                            title:
+                                Text('تعديل مكان ${street.name} على الخريطة'),
                           ),
+                          body: street.getMapView(
+                              editMode: true, useGPSIfNull: true),
                         ),
-                        child: FutureBuilder(
-                            future: street.areaId == null
-                                ? null
-                                : street.getAreaName(),
-                            builder: (con, data) {
-                              if (data.connectionState ==
-                                  ConnectionState.done) {
-                                return Text(data.data);
-                              } else if (data.connectionState ==
-                                  ConnectionState.waiting) {
-                                return LinearProgressIndicator();
-                              } else {
-                                return Text('لا يوجد');
-                              }
-                            }),
+                      ),
+                    );
+                    if (rslt == null) {
+                      street.locationPoints = oldPoints;
+                    } else if (rslt == false) {
+                      street.locationPoints = null;
+                    }
+                  }),
+              Container(
+                padding: EdgeInsets.symmetric(vertical: 4.0),
+                child: InkWell(
+                  onTap: () => _selectDate(context),
+                  child: InputDecorator(
+                    decoration: InputDecoration(
+                      labelText: 'تاريخ أخر زيارة',
+                      border: OutlineInputBorder(
+                        borderSide:
+                            BorderSide(color: Theme.of(context).primaryColor),
                       ),
                     ),
+                    child: Text(DateFormat('yyyy/M/d').format(
+                      street.lastVisit.toDate(),
+                    )),
                   ),
                 ),
-                ElevatedButton.icon(
-                    icon: Icon(
-                      const IconData(0xe568, fontFamily: 'MaterialIconsR'),
-                    ),
-                    label: Text('تعديل مكان الشارع على الخريطة'),
-                    onPressed: () async {
-                      var oldPoints = street.locationPoints?.sublist(0);
-                      var rslt = await Navigator.of(context).push(
-                        MaterialPageRoute(
-                          builder: (context) => Scaffold(
-                            appBar: AppBar(
-                              actions: <Widget>[
-                                IconButton(
-                                  icon: Icon(Icons.done),
-                                  onPressed: () => Navigator.pop(context, true),
-                                  tooltip: 'حفظ',
-                                ),
-                                IconButton(
-                                  icon: Icon(Icons.delete),
-                                  onPressed: () =>
-                                      Navigator.pop(context, false),
-                                  tooltip: 'حذف التحديد',
-                                )
-                              ],
-                              title:
-                                  Text('تعديل مكان ${street.name} على الخريطة'),
-                            ),
-                            body: street.getMapView(
-                                editMode: true, useGPSIfNull: true),
-                          ),
-                        ),
-                      );
-                      if (rslt == null) {
-                        street.locationPoints = oldPoints;
-                      } else if (rslt == false) {
-                        street.locationPoints = null;
-                      }
-                    }),
-                Container(
-                  padding: EdgeInsets.symmetric(vertical: 4.0),
-                  child: Focus(
-                    focusNode: foci[2],
-                    child: InkWell(
-                      onTap: _selectDate,
-                      child: InputDecorator(
-                        decoration: InputDecoration(
-                          labelText: 'تاريخ أخر زيارة',
-                          border: OutlineInputBorder(
-                            borderSide: BorderSide(
-                                color: Theme.of(context).primaryColor),
-                          ),
-                        ),
-                        child: Text(DateFormat('yyyy/M/d').format(
-                          street.lastVisit.toDate(),
-                        )),
+              ),
+              Container(
+                padding: EdgeInsets.symmetric(vertical: 4.0),
+                child: InkWell(
+                  onTap: () => _selectDate2(context),
+                  child: InputDecorator(
+                    decoration: InputDecoration(
+                      labelText: 'تاريخ أخر زيارة (للأب الكاهن)',
+                      border: OutlineInputBorder(
+                        borderSide:
+                            BorderSide(color: Theme.of(context).primaryColor),
                       ),
                     ),
+                    child: street.fatherLastVisit != null
+                        ? Text(DateFormat('yyyy/M/d').format(
+                            street.fatherLastVisit.toDate(),
+                          ))
+                        : Text(DateFormat('yyyy/M/d').format(
+                            DateTime(DateTime.now().year, DateTime.now().month,
+                                DateTime.now().day),
+                          )),
                   ),
                 ),
-                Container(
-                  padding: EdgeInsets.symmetric(vertical: 4.0),
-                  child: Focus(
-                    focusNode: foci[3],
-                    child: InkWell(
-                      onTap: _selectDate2,
-                      child: InputDecorator(
-                        decoration: InputDecoration(
-                          labelText: 'تاريخ أخر زيارة (للأب الكاهن)',
-                          border: OutlineInputBorder(
-                            borderSide: BorderSide(
-                                color: Theme.of(context).primaryColor),
-                          ),
-                        ),
-                        child: street.fatherLastVisit != null
-                            ? Text(DateFormat('yyyy/M/d').format(
-                                street.fatherLastVisit.toDate(),
-                              ))
-                            : Text(DateFormat('yyyy/M/d').format(
-                                DateTime(DateTime.now().year,
-                                    DateTime.now().month, DateTime.now().day),
-                              )),
-                      ),
-                    ),
-                  ),
-                ),
-                ElevatedButton.icon(
-                  style: ElevatedButton.styleFrom(primary: street.color),
-                  onPressed: selectColor,
-                  icon: Icon(Icons.color_lens),
-                  label: Text('اللون'),
-                ),
-              ],
-            ),
+              ),
+              ElevatedButton.icon(
+                style: ElevatedButton.styleFrom(primary: street.color),
+                onPressed: selectColor,
+                icon: Icon(Icons.color_lens),
+                label: Text('اللون'),
+              ),
+            ].map((w) => Focus(child: w)).toList(),
           ),
         ),
       ),
@@ -234,40 +210,50 @@ class _EditStreetState extends State<EditStreet> {
     street.lastVisit = Timestamp.fromDate(value);
   }
 
-  void delete() {
-    showDialog(
-      context: context,
-      builder: (context) => DataDialog(
-        title: Text(street.name),
-        content: Text(
-            'هل أنت متأكد من حذف ${street.name} وكل ما بداخله من عائلات وأشخاص؟'),
-        actions: <Widget>[
-          TextButton(
-            onPressed: () async {
-              ScaffoldMessenger.of(context).showSnackBar(
-                SnackBar(
-                  content: Text('جار حذف الشارع وما بداخله من بيانات...'),
-                  duration: Duration(minutes: 20),
-                ),
-              );
-              await FirebaseFirestore.instance
-                  .collection('Streets')
-                  .doc(street.id)
-                  .delete();
-              Navigator.of(context).pop();
-              Navigator.of(context).pop('deleted');
-            },
-            child: Text('نعم'),
+  void delete() async {
+    if (await showDialog(
+          context: context,
+          builder: (context) => DataDialog(
+            title: Text(street.name),
+            content: Text(
+                'هل أنت متأكد من حذف ${street.name} وكل ما بداخله من عائلات وأشخاص؟'),
+            actions: <Widget>[
+              TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop(true);
+                },
+                child: Text('نعم'),
+              ),
+              TextButton(
+                onPressed: () {
+                  Navigator.of(context).pop();
+                },
+                child: Text('تراجع'),
+              ),
+            ],
           ),
-          TextButton(
-            onPressed: () {
-              Navigator.of(context).pop();
-            },
-            child: Text('تراجع'),
-          ),
-        ],
-      ),
-    );
+        ) ==
+        true) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('جار حذف الشارع وما بداخله من بيانات...'),
+          duration: Duration(minutes: 20),
+        ),
+      );
+      if (await Connectivity().checkConnectivity() != ConnectivityResult.none)
+        await FirebaseFirestore.instance
+            .collection('Streets')
+            .doc(street.id)
+            .delete();
+      else {
+        // ignore: unawaited_futures
+        FirebaseFirestore.instance
+            .collection('Streets')
+            .doc(street.id)
+            .delete();
+      }
+      Navigator.pop(context, 'deleted');
+    }
   }
 
   @override
@@ -321,18 +307,36 @@ class _EditStreetState extends State<EditStreet> {
 
         street.lastEdit = auth.FirebaseAuth.instance.currentUser.uid;
 
+        bool update = street.id != '';
         if (street.id == '') {
-          street.id =
-              (await FirebaseFirestore.instance.collection('Streets').add(
-                        street.getMap(),
-                      ))
-                  .id;
-        } else {
+          street.id = FirebaseFirestore.instance.collection('Streets').doc().id;
+        }
+
+        if (update &&
+            await Connectivity().checkConnectivity() !=
+                ConnectivityResult.none) {
           await street.ref.update(
             street.getMap()..removeWhere((key, value) => old[key] == value),
           );
+        } else if (update) {
+          //Intentionally unawaited because of no internet connection
+          // ignore: unawaited_futures
+          street.ref.update(
+            street.getMap()..removeWhere((key, value) => old[key] == value),
+          );
+        } else if (await Connectivity().checkConnectivity() !=
+            ConnectivityResult.none) {
+          await street.ref.set(
+            street.getMap(),
+          );
+        } else {
+          //Intentionally unawaited because of no internet connection
+          // ignore: unawaited_futures
+          street.ref.set(
+            street.getMap(),
+          );
         }
-        ScaffoldMessenger.of(context).hideCurrentSnackBar();
+
         Navigator.of(context).pop(street.ref);
       } else {
         await showDialog(
@@ -392,6 +396,7 @@ class _EditStreetState extends State<EditStreet> {
                 street.areaId = (await Navigator.of(context)
                         .pushNamed('Data/EditArea')) as DocumentReference ??
                     street.areaId;
+                FocusScope.of(context).nextFocus();
                 setState(() {});
               },
               tooltip: 'إضافة منطقة جديدة',
@@ -447,13 +452,14 @@ class _EditStreetState extends State<EditStreet> {
             setState(() {
               street.color = color;
             });
+            FocusScope.of(context).nextFocus();
           },
         ),
       ),
     );
   }
 
-  Future _selectDate() async {
+  Future<void> _selectDate(BuildContext context) async {
     final DateTime picked = await showDatePicker(
       context: context,
       initialDate: street.lastVisit.toDate(),
@@ -463,10 +469,11 @@ class _EditStreetState extends State<EditStreet> {
     if (picked != null && Timestamp.fromDate(picked) != street.lastVisit)
       setState(() {
         street.lastVisit = Timestamp.fromDate(picked);
+        FocusScope.of(context).nextFocus();
       });
   }
 
-  Future _selectDate2() async {
+  Future<void> _selectDate2(BuildContext context) async {
     DateTime picked = await showDatePicker(
       context: context,
       initialDate: street.fatherLastVisit?.toDate() ?? DateTime.now(),
@@ -476,6 +483,7 @@ class _EditStreetState extends State<EditStreet> {
     if (picked != null && Timestamp.fromDate(picked) != street.fatherLastVisit)
       setState(() {
         street.fatherLastVisit = Timestamp.fromDate(picked);
+        FocusScope.of(context).nextFocus();
       });
   }
 }
